fix(sdk): Fix HMAC key length back to bits

+++ libs/sdk/assembly/crypto/crypto_subtle_idl_v1.ts
@@ -94,7 +94,7 @@ export class aes_metadata {
 @json
 export class hmac_metadata {
     sha_metadata!: sha_metadata;
-    length: u16 = 0;
+    length: u32 = 0;
 }
 
 @json
